refactor: Extract root transformation logic into its own module
